Make C API modularized - extract PWS and OTP
