+++ NK_C_API_helpers.cpp
@@ -1,0 +1,5 @@
+#include <string>
+
+void clear_string(std::string &s) {
+  std::fill(s.begin(), s.end(), ' ');
+}
